requirements.txt
discord.py
pycoingecko
prettytable
fastapi
uvicorn
Flask==2.1.1
